Extract Node-API files into a separate folder
