+++ API/hermes_node_api/node_api/CMakeLists.txt
@@ -3,12 +3,11 @@
 # This source code is licensed under the MIT license found in the
 # LICENSE file in the root directory of this source tree.
 
-add_library(hermesNodeApi INTERFACE
+add_library(nodeApi INTERFACE
   js_native_api_types.h
-  js_native_api.h
-  js_runtime_api.h)
+  js_native_api.h)
 
-target_include_directories(hermesNodeApi INTERFACE ${CMAKE_CURRENT_SOURCE_DIR})
+target_include_directories(nodeApi INTERFACE ${CMAKE_CURRENT_SOURCE_DIR})
 
-install(DIRECTORY "${PROJECT_SOURCE_DIR}/API/node-api/" DESTINATION include
+install(DIRECTORY "${PROJECT_SOURCE_DIR}/API/node_api/" DESTINATION include
   FILES_MATCHING PATTERN "*.h")
